requirements.txt
tensorflow
gradio
matplotlib
Pillow

